refactor: 테스트 더블 디렉터리 이동
