+++ app/src/test/java/woowacourse/shopping/viewmodel/fakes/FakeRecentlyViewedProductRepository.kt
@@ -7,14 +7,14 @@ import woowacourse.shopping.data.local.repository.RecentlyViewedProductRepositor
 import woowacourse.shopping.domain.Product
 
 class FakeRecentlyViewedProductRepository : RecentlyViewedProductRepository {
-    private val _history = MutableStateFlow<List<RecentlyViewedProductEntity>>(emptyList())
+    private val _history = MutableStateFlow<List<Long>>(emptyList())
 
-    override fun getAll(): Flow<List<RecentlyViewedProductEntity>?> = _history
+    override fun getAll(): Flow<List<Long>?> = _history
 
     override suspend fun updateList(product: Product) {
         val current = _history.value.toMutableList()
-        current.removeAll { it.id == product.id }
-        current.add(0, RecentlyViewedProductEntity(product.id))
+        current.removeAll { it == product.id }
+        current.add(0, product.id)
         if (current.size > 10) {
             _history.value = current.take(10)
         } else {
@@ -22,5 +22,5 @@ class FakeRecentlyViewedProductRepository : RecentlyViewedProductRepository {
         }
     }
 
-    override fun getLatestItem(): Flow<Long?> = MutableStateFlow(_history.value.firstOrNull()?.id)
+    override fun getLatestItem(): Flow<Long?> = MutableStateFlow(_history.value.firstOrNull())
 }
