+++ app/src/test/java/woowacourse/shopping/viewmodel/fakes/server/MockProducts.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.mock
+package woowacourse.shopping.viewmodel.fakes.server
 
 object MockProducts {
     val products: String =
